requirements.txt
# ============================================================
# Thing Identification System - Full Dependency List
# Platform : Windows (CPU only)
# Python   : 3.11
# ============================================================
#
# SETUP STEPS:
#
# [On internet machine]
# 1. Install Python 3.11: https://www.python.org/downloads/release/python-3119/
# 2. Create venv:
#       py -3.11 -m venv venv
#       venv\Scripts\activate
# 3. Download packages (this file):
#       pip download -r requirements.txt -d ./offline_packages
# 4. Download PyTorch CPU separately:
#       pip download torch==2.6.0 torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cpu -d ./offline_packages
#
# [On offline local machine]
# 5. Install Python 3.11 (same version)
# 6. Create venv:
#       py -3.11 -m venv venv
#       venv\Scripts\activate
# 7. Install all packages offline:
#       pip install --no-index --find-links=./offline_packages -r requirements.txt
#
# ============================================================

# FastAPI + dependencies
fastapi==0.110.0
starlette==0.36.3
pydantic==2.6.4
pydantic-core==2.16.3
annotated-types==0.6.0
typing_extensions==4.11.0

# Uvicorn - uvloop and httptools excluded (not supported on Windows)
uvicorn==0.29.0
websockets==12.0
python-dotenv==1.0.1
watchfiles==0.21.0

# Async support
anyio==4.3.0
sniffio==1.3.1
idna==3.7

# File upload
python-multipart==0.0.9

# HTTP client
httpx==0.27.0
httpcore==1.0.5
certifi==2024.2.2
h11==0.14.0

# NOTE: torch and torchvision are NOT listed here.
# Download them separately using Command 2 below (pytorch.org index).
# They will be placed in the same ./offline_packages folder.

# Required by torch and torchvision
numpy==1.26.4
Pillow==10.3.0
requests==2.31.0
charset-normalizer==3.3.2
urllib3==2.2.1
filelock==3.13.4
fsspec==2024.3.1
networkx==3.3
jinja2==3.1.3
MarkupSafe==2.1.5
sympy==1.12
mpmath==1.3.0

# Utilities
packaging==24.0
six==1.16.0
